fixed_files

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -11,7 +11,7 @@
 list_t *add_node(list_t **head, const char *str)
 {
 	list_t *new_l;
-	unsigned int lenght = 0;
+	long int lenght = 0; /**a whole integer whose range is greater than or equal to that of a standard integer on the same machine**/
 
 	new_l = malloc(sizeof(list_t));
 	if (!new_l)
